wip: debugging 64 bit invalidation cube

+++ ext/skeletontricks/skeletontricks.pyx
@@ -59,11 +59,11 @@ ctypedef fused INTEGER:
   unsigned char
 
 cdef extern from "skeletontricks.hpp" namespace "skeletontricks":
-  cdef int _roll_invalidation_cube(
+  cdef size_t _roll_invalidation_cube(
     uint8_t* labels, float* DBF,
-    int sx, int sy, int sz,
+    size_t sx, size_t sy, size_t sz,
     float wx, float wy, float wz,
-    int* path, int path_size,
+    size_t* path, size_t path_size,
     float scale, float constant
   )
 
@@ -727,12 +727,12 @@ def roll_invalidation_cube(
   in a cube around each vertex. In contrast to `roll_invalidation_ball`,
   this function runs in time linear in the number of image pixels.
   """
-  cdef int sx, sy, sz 
+  cdef size_t sx, sy, sz 
   sx = labels.shape[0]
   sy = labels.shape[1]
   sz = labels.shape[2]
 
-  cdef int sxy = sx * sy
+  cdef size_t sxy = sx * sy
 
   cdef float wx, wy, wz
   (wx, wy, wz) = anisotropy
@@ -741,15 +741,15 @@ def roll_invalidation_cube(
     coord[0] + sx * coord[1] + sxy * coord[2] 
     for coord in path if tuple(coord) not in invalid_vertices 
   ]
-  path = np.array(path, dtype=np.int32)
+  path = np.array(path, dtype=np.uint64)
 
-  cdef int[:] pathview = path
+  cdef size_t[:] pathview = path
 
   cdef int invalidated = _roll_invalidation_cube(
     <uint8_t*>&labels[0,0,0], <float*>&DBF[0,0,0],
     sx, sy, sz, 
     wx, wy, wz,
-    <int*>&pathview[0], path.size,
+    <size_t*>&pathview[0], path.size,
     scale, const
   )
